Factor out settings classes
